Replace bare block serialization with message serialization
It is not particularly useful to serialize a block to JSON.
When a node sends a message to another node could include a block, but also some "header".
Now a message consists of a message type, a sender and an optional payload (which could be a block or a blockchain).

These serialization functions are in Naivecoin::Core library, but they might be moved later, since strictly speaking they belong to a node.